requirements.txt
sync2asyncio
oss2
pandas
aerich==0.5.3
aioboto3==7.0.0
aiobotocore==1.3.1
aiofiles==0.7.0
aiohttp==3.7.4.post0
aioitertools==0.7.1
aiomysql==0.0.21
aioredis==1.3.1
aiosqlite==0.16.1
alembic==1.6.5
aliyun-python-sdk-core==2.13.35
amqp==5.0.6
anyio==3.2.0
APScheduler==3.7.0
asgiref==3.3.4
async-timeout==3.0.1
atomicwrites==1.4.0
attrs==21.2.0
Authlib==0.15.4
autopep8==1.5.7
Babel==2.9.1
bcrypt==3.2.0
billiard==3.6.4.0
boto3==1.17.49
botocore==1.20.49
bracex==2.1.1
casbin==1.1.2
casbin-sqlalchemy-adapter==0.3.1
cchardet==2.1.7
celery==5.0.5
certifi==2021.5.30
cffi==1.14.5
chardet==4.0.0
click==7.1.2
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.2.0
colorama==0.4.4
cryptography==3.4.7
ddlparse==1.9.0
dictdiffer==0.8.1
dnspython==2.1.0
ecdsa==0.17.0
email-validator==1.1.2
fastapi==0.65.2
fastapi-admin==1.0.1
greenlet==1.1.0
gunicorn==20.1.0
h11==0.12.0
hiredis==2.0.0
httpcore==0.13.6
httpx==0.18.2
idna==2.10
iniconfig==1.1.1
iso8601==0.1.14
itsdangerous==2.0.1
Jinja2==3.0.1
jmespath==0.10.0
kombu==5.1.0
limits==1.5.1
loguru==0.5.3
lxml==4.6.3
Mako==1.1.4
MarkupSafe==2.0.1
multidict==5.1.0
optionaldict==0.1.2
packaging==20.9
passlib==1.7.4
pendulum==2.1.2
pluggy==0.13.1
prompt-toolkit==3.0.18
py==1.10.0
pyasn1==0.4.8
pycodestyle==2.7.0
pycparser==2.20
pydantic==1.8.2
PyMySQL==0.9.3
pyOpenSSL==19.1.0
pyparsing==2.4.7
pypika-tortoise==0.1.1
pytest==6.2.4
python-dateutil==2.8.1
python-dotenv==0.17.1
python-editor==1.0.4
python-jose==3.3.0
python-multipart==0.0.5
pytz==2021.1
pytzdata==2020.1
PyYAML==5.4.1
qcloudsms-py==0.1.4
redis==3.5.3
requests==2.25.1
rfc3986==1.5.0
rsa==4.7.2
s3transfer==0.3.7
six==1.16.0
slowapi==0.1.4
sniffio==1.2.0
SQLAlchemy==1.4.20
starlette==0.14.2
toml==0.10.2
tortoise-orm==0.17.4
typing-extensions==3.10.0.0
tzlocal==2.1
urllib3==1.26.5
uvicorn==0.14.0
vine==5.0.0
watchgod==0.7
wcmatch==8.2
wcwidth==0.2.5
websockets==9.1
wechatpy==1.8.15
weixin-python==0.5.7
win32-setctime==1.0.3
wrapt==1.12.1
xmltodict==0.12.0
yarl==1.6.3
casbin_tortoise_adapter